Changed expense amount type to BigDecimal.

+++ expense-tracker-server/src/main/java/inzagher/expense/tracker/server/dto/CategoryReportItemDTO.java
@@ -1,5 +1,6 @@
 package inzagher.expense.tracker.server.dto;
 
+import java.math.BigDecimal;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlType;
@@ -8,7 +9,7 @@ import javax.xml.bind.annotation.XmlType;
 @XmlAccessorType(XmlAccessType.FIELD)
 public class CategoryReportItemDTO {
     private CategoryDTO category;
-    private Float amount;
+    private BigDecimal amount;
 
     public CategoryDTO getCategory() {
         return category;
@@ -18,11 +19,11 @@ public class CategoryReportItemDTO {
         this.category = category;
     }
 
-    public Float getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Float amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
 }
